% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\alias{roxy_tag_parse.roxy_tag_metadata}
\alias{roxy_tag_rd.roxy_tag_metadata}
\alias{format.rd_section_metadata}
\alias{roclet_process.roclet_metadata}
\alias{roclet_output.roclet_metadata}
\title{roclet to parse @metadata tag}
\usage{
metadata()

\method{roxy_tag_parse}{roxy_tag_metadata}(x)

\method{roxy_tag_rd}{roxy_tag_metadata}(x, base_path, env)

\method{format}{rd_section_metadata}(x, ...)

\method{roclet_process}{roclet_metadata}(x, blocks, env, base_path)

\method{roclet_output}{roclet_metadata}(x, results, base_path, ...)
}
\description{
roclet to parse @metadata tag
}
